tests and stuff

+++ lib/SchemaBuilder.ts
@@ -1,4 +1,5 @@
 import {Model} from "objection";
+import CustomModel, {SchemaDefinition} from "./CustomModel";
 // import {makeExecutableSchema} from "@graphql-tools/schema";
 
 // type CustomResolver = { [index in string]?: any; }
@@ -6,9 +7,10 @@ type CustomResolver = any
 type Options = {
     mutationOptions?: MutationOptions, queryOptions?: QueryOptions, extendQueries?: string,
     extendMutations?: string, extendTypes?: string, extendResolvers?: CustomResolver,
-    extendMutationResolvers?: CustomResolver, extendQueryResolvers?: CustomResolver
+    extendMutationResolvers?: CustomResolver, extendQueryResolvers?: CustomResolver,
+    defaultProperties?: { [index: string]: SchemaDefinition }
 }
-type ModelsObject = { [index: string]: typeof Model }
+type ModelsObject = { [index: string]: typeof CustomModel }
 type ArgumentKeys =
     'eq' | 'gt' | 'gte' | 'lt' | 'lte' | 'like' | 'isNull' | 'likeNoCase' | 'in' | 'notIn' | 'orderBy'
     | 'orderByDesc' | 'range' | 'limit' | 'offset'
@@ -16,7 +18,6 @@ type ArgumentType = string | false
 type QueryOptions = { [index in ArgumentKeys]?: ArgumentType; } | boolean;
 type MutationOptions = { create: boolean, update: boolean, delete: boolean } | boolean
 
-//todo custommodel
 class SchemaBuilder {
     models: ModelsObject = {}
     queryOptions: QueryOptions = true
@@ -27,9 +28,10 @@ class SchemaBuilder {
     extendResolvers: CustomResolver = {}
     extendMutationResolvers: CustomResolver = {}
     extendQueryResolvers: CustomResolver = {}
+    defaultProperties: { [p: string]: SchemaDefinition } = {}
 
 
-    constructor(models: (typeof Model)[], options?: Options) {
+    constructor(models: (typeof CustomModel)[], options?: Options) {
         this.models = models.reduce<ModelsObject>((models, model) => ({
             ...models,
             [model.name]: model
@@ -43,12 +45,18 @@ class SchemaBuilder {
         if (options.extendResolvers) this.extendResolvers = options.extendResolvers
         if (options.extendMutationResolvers) this.extendMutationResolvers = options.extendMutationResolvers
         if (options.extendQueryResolvers) this.extendQueryResolvers = options.extendQueryResolvers
+        if (options.defaultProperties) this.defaultProperties = options.defaultProperties
+    }
+
+    getPrimaryField(model: typeof CustomModel) {
+        return model.graphqlOptions?.primaryField || 'id'
     }
 
     buildModelTypeDefs() {
         const resolveType = (type: string) => {
             if (type === 'string') return 'String'
             if (type === 'integer') return 'Int'
+            else return type.charAt(0).toUpperCase() + type.substring(1)
         }
 
 
@@ -64,33 +72,42 @@ input ${model.name}UpdateInput {`
 
             const fields = model.jsonSchema.properties
             if (!fields) throw new Error(`${model.name} jsonSchema properties is missing`)
+            const fields2: typeof fields = this.defaultProperties
+            const ff: typeof fields[] = []
+            if (fields) ff.push(fields)
+            if (fields2) ff.push(fields2)
 
-            Object.keys(fields).forEach(key => {
-                const field: { relationName: string, type: string } = fields[key] as any
-                const fieldType = field.type
-                if (fieldType === 'object') {
-                    const relationMappings: any = model.relationMappings
-                    const relation = relationMappings[key] || relationMappings[field.relationName]
-                    const relatedModel = relation.connectModel || relation.modelClass
-                    const relatedModelName = relatedModel.name
-                    const relationType = relation.relation
-                    if (relationType === Model.HasManyRelation) {
-                        modelTypeDef += `
+            console.log(ff)
+            ff.forEach(f => {
+                Object.keys(f).forEach(key => {
+                    console.log(key)
+                    const field: { relationName: string, type: string } = fields[key] as any || fields2[key]
+                    const fieldType = field.type
+                    if (fieldType === 'object') {
+                        const relationMappings: any = model.relationMappings
+                        const relation = relationMappings[key] || relationMappings[field.relationName]
+                        const relatedModel = relation.connectModel || relation.modelClass
+                        const relatedModelName = relatedModel.name
+                        const relationType = relation.relation
+                        if (relationType === Model.HasManyRelation) {
+                            modelTypeDef += `
 ${key}: [${relatedModelName}]`
+                        } else {
+                            modelTypeDef += `
+${key}: ${relatedModelName}`
+                        }
                     } else {
                         modelTypeDef += `
-${key}: ${relatedModelName}`
+${key}: ${resolveType(fieldType)}`
+                        modelTypeDefCreate += `
+${key}: ${resolveType(fieldType)}`
+                        modelTypeDefUpdate += `
+${key}: ${resolveType(fieldType)}`
                     }
-                } else {
-                    modelTypeDef += `
-${key}: ${resolveType(fieldType)}`
-                    modelTypeDefCreate += `
-${key}: ${resolveType(fieldType)}`
-                    modelTypeDefUpdate += `
-${key}: ${resolveType(fieldType)}`
-                }
+                })
 
             })
+
             modelTypeDefs += modelTypeDef + '\n}' + modelTypeDefCreate + '\n}' + modelTypeDefUpdate + '\n}\n' + this.extendTypes
         }
         return modelTypeDefs
@@ -103,9 +120,11 @@ ${key}: ${resolveType(fieldType)}`
         for (const key of Object.keys(this.models)) {
             const model = this.models[key]
             const camelCaseName = `${model.name.charAt(0).toLowerCase()}${model.name.substring(1)}`
-            queryTypeDefs += `
-${camelCaseName}(id: Int!): ${model.name}
-${camelCaseName}s(idArray: [Int]): [${model.name}]
+            const idField = this.getPrimaryField(model)
+            if (model.graphqlOptions?.generateDefaultQueries)
+                queryTypeDefs += `
+${camelCaseName}(${idField}: Int!): ${model.name}
+${camelCaseName}s(${idField}Array: [Int]): [${model.name}]
 `
         }
         queryTypeDefs += this.extendQueries + '\n}'
@@ -118,10 +137,12 @@ ${camelCaseName}s(idArray: [Int]): [${model.name}]
         for (const key of Object.keys(this.models)) {
             const model = this.models[key]
             const camelCaseName = `${model.name.charAt(0).toLowerCase()}${model.name.substring(1)}`
-            mutationTypeDefs += `
+            const idField = this.getPrimaryField(model)
+            if (model.graphqlOptions?.generateDefaultMutations)
+                mutationTypeDefs += `
 create${model.name}(${camelCaseName}: ${model.name}CreateInput!): ${model.name}!
-update${model.name}(id: Int!, ${camelCaseName}: ${model.name}UpdateInput!): ${model.name}!
-delete${model.name}(id: Int!): Boolean
+update${model.name}(${idField}: Int!, ${camelCaseName}: ${model.name}UpdateInput!): ${model.name}!
+delete${model.name}(${idField}: Int!): Boolean
 `
         }
         mutationTypeDefs += this.extendMutations + '\n}'
@@ -133,39 +154,47 @@ delete${model.name}(id: Int!): Boolean
     }
 
     buildQueryResolvers() {
-        const queryResolvers: any = {}
+        let queryResolvers: any = {}
         for (const key of Object.keys(this.models)) {
             const model = this.models[key]
             const camelCaseName = `${model.name.charAt(0).toLowerCase()}${model.name.substring(1)}`
-
-            queryResolvers[camelCaseName] = async (root: any, args: any, context: any, info: any) => {
-                return model.query().findById(args.id).throwIfNotFound().select()
-            }
-            queryResolvers[`${camelCaseName}s`] =
-                async (root: any, args: any, context: any, info: any) => {
-                    const {idArray} = args
-                    const queryBuilder = model.query().select()
-                    if (idArray) queryBuilder.findByIds(args.idArray as number[])
-                    return queryBuilder
+            if (model.graphqlOptions?.generateDefaultQueries) {
+                queryResolvers[camelCaseName] = async (root: any, args: any, context: any, info: any) => {
+                    return model.query().findById(args.id).throwIfNotFound().select()
                 }
+                queryResolvers[`${camelCaseName}s`] =
+                    async (root: any, args: any, context: any, info: any) => {
+                        const {idArray} = args
+                        const queryBuilder = model.query().select()
+                        if (idArray) queryBuilder.findByIds(args.idArray as number[])
+                        return queryBuilder
+                    }
+            }
+
+            if (model.graphqlOptions && model.graphqlOptions.queryResolvers) queryResolvers = {...queryResolvers, ...model.graphqlOptions.queryResolvers}
+
         }
         return {...queryResolvers, ...this.extendQueryResolvers}
     }
 
     buildMutationResolvers() {
-        const mutationResolvers: any = {}
+        let mutationResolvers: any = {}
         for (const key of Object.keys(this.models)) {
             const model = this.models[key]
 
-            mutationResolvers[`create${model.name}`] = async (root: any, args: any, context: any, info: any) => {
-                return model.query().insert(args.role).returning('*')
+            if (model.graphqlOptions?.generateDefaultMutations) {
+                mutationResolvers[`create${model.name}`] = async (root: any, args: any, context: any, info: any) => {
+                    return model.query().insert(args.role).returning('*')
+                }
+                mutationResolvers[`update${model.name}`] = async (root: any, args: any, context: any, info: any) => {
+                    return model.query().patchAndFetchById(args.id, args.role).returning('*')
+                }
+                mutationResolvers[`delete${model.name}`] = async (root: any, args: any, context: any) => {
+                    return Boolean(await model.query().deleteById(args.id))
+                }
             }
-            mutationResolvers[`update${model.name}`] = async (root: any, args: any, context: any, info: any) => {
-                return model.query().patchAndFetchById(args.id, args.role).returning('*')
-            }
-            mutationResolvers[`delete${model.name}`] = async (root: any, args: any, context: any) => {
-                return Boolean(await model.query().deleteById(args.id))
-            }
+
+            if (model.graphqlOptions && model.graphqlOptions.mutationResolvers) mutationResolvers = {...mutationResolvers, ...model.graphqlOptions.mutationResolvers}
         }
         return {...mutationResolvers, ...this.extendMutationResolvers}
     }
@@ -181,22 +210,25 @@ delete${model.name}(id: Int!): Boolean
                     const relation = b[a]
                     const relationName = b[a].graphqlName || a
                     if (!relation.graphql) return
-                    switch (relation.connectType) {
-                        case "one":
+                    console.dir(relation.relation, {depth: null})
+                    switch (relation.relation) {
+                        case Model.HasOneRelation:
+                        case Model.BelongsToOneRelation:
                             resolver[relationName] = async (parent: any, args: any, context: any, info: any) => {
                                 const res = await model.relatedQuery(a).for(parent.id!).select()
                                 return res[0]
                             }
                             break;
-                        case "many":
-                            resolver[relationName] = async (parent: any, args: any, context: any, info: any) => {
-                                return model.relatedQuery(a).for(parent.id!).select()
-                            }
-                            break;
-                        case "through":
-                            resolver[relationName] = async (parent: any, args: any, context: any, info: any) => {
-                                const r = await model.relatedQuery(relation.connectQuery).for(parent.id!).select('id')
-                                return relation.modelClass.relatedQuery(a).for(r.map((cr: any) => cr.id!)).select()
+                        case Model.HasManyRelation:
+                            if (Object.keys(relation.join).includes('through')) {
+                                resolver[relationName] = async (parent: any, args: any, context: any, info: any) => {
+                                    const r = await model.relatedQuery(relation.connectQuery).for(parent.id!).select('id')
+                                    return relation.modelClass.relatedQuery(a).for(r.map((cr: any) => cr.id!)).select()
+                                }
+                            } else {
+                                resolver[relationName] = async (parent: any, args: any, context: any, info: any) => {
+                                    return model.relatedQuery(a).for(parent.id!).select()
+                                }
                             }
                             break;
                     }
